project/api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pdfminer.six==20231228
spacy==3.7.4
python-dotenv==1.0.1
sentence-transformers==2.5.1
alembic==1.13.1